PAYS-85: Refactor the handlers
