--- custom/paystack/infrastructure/event/src/main/java/com/paystack/fineract/infrastructure/event/external/service/command/KafkaNotificationCommandHandler.java
@@ -1,55 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements. See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership. The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package com.paystack.fineract.infrastructure.event.external.service.command;
-
-import org.apache.fineract.infrastructure.core.api.JsonCommand;
-import org.apache.fineract.infrastructure.core.data.CommandProcessingResult;
-
-/**
- * Command handler interface for Kafka notification operations.
- */
-public interface KafkaNotificationCommandHandler {
-
-    /**
-     * Handle retry failed notification command.
-     *
-     * @param command
-     *            the JSON command
-     * @return CommandProcessingResult
-     */
-    CommandProcessingResult retryFailedNotification(JsonCommand command);
-
-    /**
-     * Handle retry failed notifications by account ID command.
-     *
-     * @param command
-     *            the JSON command
-     * @return CommandProcessingResult
-     */
-    CommandProcessingResult retryFailedNotificationsByAccountId(JsonCommand command);
-
-    /**
-     * Handle retry all failed notifications command.
-     *
-     * @param command
-     *            the JSON command
-     * @return CommandProcessingResult
-     */
-    CommandProcessingResult retryAllFailedNotifications(JsonCommand command);
-}
